Updated empty_cells
The first code was an attempt to read the .txt file line by line to find any empty cells in any of the columns. Even though the code fetched header row correctly it was unable to recognize any empty cells in any of the columns. Changed the code to use awk script in order to better read all the rows based on seperator